+++ quantify_scheduler/backends/types/qblox.py
@@ -544,8 +544,6 @@ class QbloxBaseDescription(HardwareDescription):
     """The reference source for the instrument."""
     sequence_to_file: bool = False
     """Write sequencer programs to files for (all modules in this) instrument."""
-    align_qasm_fields: bool = False
-    """If True, make QASM program more human-readable by aligning its fields."""
 
 
 class ComplexChannelDescription(DataStructure):